refactor : move ordertype to eatin, separate menu and menugroup
