+++ src/main/java/kitchenpos/common/name/infra/PurgomalumClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.common.infra;
+package kitchenpos.common.name.infra;
 
 public interface PurgomalumClient {
     boolean containsProfanity(String text);
