+++ src/main/java/kitchenpos/eatin/domain/Order.java
@@ -1,7 +1,5 @@
 package kitchenpos.eatin.domain;
 
-import kitchenpos.common.code.OrderType;
-
 import javax.persistence.*;
 import java.time.LocalDateTime;
 import java.util.List;
