+++ src/main/java/kitchenpos/menu/application/MenuService.java
@@ -1,7 +1,9 @@
 package kitchenpos.menu.application;
 
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.common.name.infra.PurgomalumClient;
 import kitchenpos.menu.domain.*;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.springframework.stereotype.Service;
