+++ src/main/java/kitchenpos/menugroup/infra/JpaMenuGroupRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.menu.infra;
+package kitchenpos.menugroup.infra;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
