+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -3,9 +3,9 @@ package kitchenpos.menu.application;
 import kitchenpos.common.infra.FakePurgomalumClient;
 import kitchenpos.menu.infra.InMemoryMenuGroupRepository;
 import kitchenpos.menu.infra.InMemoryMenuRepository;
-import kitchenpos.common.infra.PurgomalumClient;
+import kitchenpos.common.name.infra.PurgomalumClient;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuProduct;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
