+++ src/test/java/kitchenpos/menu/infra/InMemoryMenuGroupRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.menu.infra;
 
-import kitchenpos.menu.domain.MenuGroup;
-import kitchenpos.menu.domain.MenuGroupRepository;
+import kitchenpos.menugroup.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroupRepository;
 
 import java.util.*;
